+++ ui/src/elle/components/InfoElement.js
@@ -1,16 +1,46 @@
 import React, {useState} from 'react';
 import './styles/InfoElement.css'
 import {Box} from "@mui/material";
-import word_analyzer_video from '../resources/videos/word_analyzer_full.webm'
 import {useTranslation} from "react-i18next";
-import PlayCircleIcon from '@mui/icons-material/PlayCircle';
 import InfoElementTabCard from "./InfoElementTabCard";
 import {ElleOuterDivStyle} from "../utils/constants";
+import i18n from "i18next";
+import wordlistImgEt from "../resources/images/home/sonaloend_et.png";
+import wordlistImgEn from "../resources/images/home/sonaloend_en.png";
+import wordcontextImgEt from "../resources/images/home/sonakontekstis_et.png";
+import wordcontextImgEn from "../resources/images/home/sonakontekstis_en.png";
+import collocateImgEt from "../resources/images/home/naabersonad_et.png";
+import collocateImgEn from "../resources/images/home/naabersonad_en.png";
+import analyserImgEt from "../resources/images/home/sonaanalyys_et.png";
+import analyserImgEn from "../resources/images/home/sonaanalyys_en.png";
+import textRaterImg from "../resources/images/home/tekstihindaja.png"
+import textQueryImg from "../resources/images/home/paring.png"
+import wordPatternImg from "../resources/images/home/mustrileidja.png"
 
 const InfoElement = () => {
 
   const {t} = useTranslation();
   const [openTab, setOpentab] = useState("corrector-tab-card");
+  const [wordlistImg, setWordlistImg] = useState(i18n.language === 'ET' ? wordlistImgEt : wordlistImgEn);
+  const [wordcontextImg, setWordcontextImg] = useState(i18n.language === 'ET' ? wordcontextImgEt : wordcontextImgEn);
+  const [collocateImg, setCollocateImg] = useState(i18n.language === 'ET' ? collocateImgEt : collocateImgEn);
+  const [analyserImg, setAnalyserImg] = useState(i18n.language === 'ET' ? analyserImgEt : analyserImgEn);
+  const [imageSelected, setImageSelected] = useState(1);
+  const images = [textRaterImg, textQueryImg, wordlistImg, wordcontextImg, collocateImg, wordPatternImg, analyserImg];
+
+  i18n.on('languageChanged', () => {
+    if (i18n.language === 'ET') {
+      setWordlistImg(wordlistImgEt);
+      setWordcontextImg(wordcontextImgEt);
+      setCollocateImg(collocateImgEt);
+      setAnalyserImg(analyserImgEt);
+    } else {
+      setWordlistImg(wordlistImgEn);
+      setWordcontextImg(wordcontextImgEn);
+      setCollocateImg(collocateImgEn);
+      setAnalyserImg(analyserImgEn);
+    }
+  });
 
   const descriptionText = {
     1: <p><b>{t('homepage_box_corrector_title')}</b> {t('homepage_box_corrector_content')}</p>,
@@ -36,14 +66,22 @@ const InfoElement = () => {
             toolInnerId={"corrector-info-popup"}
             description={descriptionText[1]}
             setOpenTab={setOpentab}
+            setImageSelected={setImageSelected}
+            imageNo={1}
+            tabReference={null}
+            linkTo={'/corrector'}
           />
           <InfoElementTabCard
-            toolID={"tekstiparing-btn"}
-            toolTitile={"Tekstipäring"}
+            toolID={"text-query-btn"}
+            toolTitile={t('common_query')}
             tabOpen={openTab}
-            toolInnerId={"tekstiparing-btn-box"}
+            toolInnerId={"text-query-btn-box"}
             description={descriptionText[2]}
             setOpenTab={setOpentab}
+            setImageSelected={setImageSelected}
+            imageNo={2}
+            tabReference={null}
+            linkTo={'/tools'}
           />
           <InfoElementTabCard
             toolID={"wordlist-tab-card"}
@@ -52,6 +90,10 @@ const InfoElement = () => {
             tabOpen={openTab}
             description={descriptionText[3]}
             setOpenTab={setOpentab}
+            setImageSelected={setImageSelected}
+            imageNo={3}
+            tabReference={'1'}
+            linkTo={'/tools'}
           />
           <InfoElementTabCard
             toolID={"word-context-tab-card"}
@@ -60,6 +102,10 @@ const InfoElement = () => {
             toolInnerId={"word-context-info-popup"}
             description={descriptionText[4]}
             setOpenTab={setOpentab}
+            setImageSelected={setImageSelected}
+            imageNo={4}
+            tabReference={'2'}
+            linkTo={'/tools'}
           />
           <InfoElementTabCard
             toolID={"neighbour-word-tab-card"}
@@ -68,6 +114,10 @@ const InfoElement = () => {
             toolInnerId={"neighbour-word-info-popup"}
             description={descriptionText[5]}
             setOpenTab={setOpentab}
+            setImageSelected={setImageSelected}
+            imageNo={5}
+            tabReference={'3'}
+            linkTo={'/tools'}
           />
           <InfoElementTabCard
             toolID={"cluster-finder-tab-card"}
@@ -76,18 +126,28 @@ const InfoElement = () => {
             toolInnerId={"cluster-finder-info-popup"}
             description={descriptionText[6]}
             setOpenTab={setOpentab}
+            setImageSelected={setImageSelected}
+            imageNo={6}
+            tabReference={'5'}
+            linkTo={'/tools'}
           />
           <InfoElementTabCard
             toolID={"word-analyser-tab-card"}
-            toolTitile={"Sõnaanalüsaator"}
+            toolTitile={t('common_word_analyser')}
             tabOpen={openTab}
             toolInnerId={"word-analyser-info-popup"}
             description={descriptionText[7]}
             setOpenTab={setOpentab}
+            setImageSelected={setImageSelected}
+            imageNo={7}
+            tabReference={'4'}
+            linkTo={'/tools'}
           />
         </Box>
+        {/*TODO videod tööle panna, kui videod valmis*/}
         <Box className="info-video">
-          <video
+          <img className="info-image" src={images[imageSelected - 1]} alt="Tööriista pilt"/>
+          {/*  <video
             className={"info-video-styling w-100"}
             controls
             src={word_analyzer_video}
@@ -99,7 +159,7 @@ const InfoElement = () => {
             <Box className="position-relative font-weight-bold elle-medium-text tools-video-text">
               <p className="text-uppercase">{t('tools_video_text')}</p>
             </Box>
-          </Box>
+          </Box>*/}
         </Box>
       </Box>
     </Box>
