+++ ui/src/elle/components/styles/DonateText.css
@@ -6,7 +6,7 @@
     z-index: 100;
     height: 60px !important;
     display: flex !important;
-    justify-content: start !important;
+    justify-content: flex-start !important;
     align-items: center !important;
     padding-left: 15px !important;
     position: fixed !important;
@@ -17,13 +17,17 @@
     transition: 0.5s !important;
 }
 
-
 .donate-text-button:hover {
-    left: calc(100vw - 310px);
+    left: calc(100vw - 260px);
     transition: 0.5s !important;
 }
 
 .donate-text-font {
     font-weight: bold;
     font-size: 1rem;
+}
+
+.donate-text-icon {
+    font-size: 40px !important;
+    margin-right: 10px !important;
 }
