+++ ui/src/elle/pages/query/Query.js
@@ -1,27 +1,19 @@
-import React, { useEffect, useState } from 'react';
+import React, {useEffect, useState} from 'react';
 import {
-  Accordion,
-  AccordionDetails,
-  AccordionSummary,
   Alert,
   Box,
   Button,
   Checkbox,
-  Chip,
   FormControl,
-  IconButton,
   InputLabel,
   ListItemIcon,
   ListItemText,
   MenuItem,
-  Modal,
   Select,
-  Tooltip,
-  Typography
+  Tooltip
 } from '@mui/material';
 import '../styles/Query.css';
 import {
-  AccordionStyle,
   addedYearOptions,
   ageOptions,
   charactersOptions,
@@ -32,7 +24,6 @@ import {
   genderOptions,
   languageOptions,
   MenuProps,
-  modalStyle,
   nationalityOptions,
   sentencesOptions,
   studyLevelOptions,
@@ -45,13 +36,13 @@ import {
   wordsOptions
 } from '../../utils/constants';
 import QueryResults from './QueryResults';
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import { useLocation, useNavigate, useSearchParams } from 'react-router-dom';
+import {useLocation, useNavigate, useSearchParams} from 'react-router-dom';
 import TextUpload from '../../components/TextUpload';
-import CloseIcon from '@mui/icons-material/Close';
-import { queryStore } from '../../store/QueryStore';
-import { loadFetch } from '../../service/LoadFetch';
-import { useTranslation } from 'react-i18next';
+import {queryStore} from '../../store/QueryStore';
+import {loadFetch} from '../../service/LoadFetch';
+import {useTranslation} from 'react-i18next';
+import ManageSearchIcon from '@mui/icons-material/ManageSearch';
+import ReadMoreIcon from '@mui/icons-material/ReadMore';
 
 export default function Query() {
 
@@ -88,6 +79,9 @@ export default function Query() {
     cZjHWUPtD: false,
     cwUSEqQLt: false
   });
+  const [filterOpen, setFilterOpen] = React.useState(true);
+  const [inputOpen, setInputOpen] = React.useState(true);
+  const [queryAnswer, setQueryAnswer] = React.useState(false);
   const [singlePropertyData, setSinglePropertyData] = useState({
     language: 'eesti',
     level: '',
@@ -214,7 +208,7 @@ export default function Query() {
             setResults([]);
           }
           setExpanded(false);
-        });
+        }).then(setQueryAnswer(true));
     }
   };
 
@@ -397,25 +391,684 @@ export default function Query() {
     }
   };
 
+  {/*TODO Query buttons need to be redone*/
+  }
+  const setFilterBoxClass = () => {
+    if (filterOpen) {
+      setInputOpen(true)
+      setFilterOpen(false)
+      document.getElementById("choose-input-button").classList.remove("button-box-open")
+      document.getElementById("choose-text-button").classList.add("button-box-open")
+    } else if (!filterOpen) {
+      setFilterOpen(true)
+      document.activeElement.blur()
+      document.getElementById("choose-text-button").classList.remove("button-box-open")
+    }
+  }
+
+  const closeInputHidden = () => {
+    if (inputOpen) {
+      setInputHidden()
+    } else if (filterOpen) {
+      setFilterBoxClass()
+    }
+  }
+
+  {/*TODO Query buttons need to be redone*/
+  }
+  const setInputHidden = () => {
+    if (inputOpen) {
+      setFilterOpen(true)
+      setInputOpen(false)
+      document.getElementById("choose-text-button").classList.remove("button-box-open")
+      document.getElementById("choose-input-button").classList.add("button-box-open")
+    } else if (!inputOpen) {
+      setInputOpen(true)
+      document.activeElement.blur()
+      document.getElementById("choose-input-button").classList.remove("button-box-open")
+    }
+  }
+
   return (
     <div>
-      <Alert severity="info">{t('tools_infobox')}</Alert>
-      <br/>
+      {/*<Alert severity="info">{t('tools_infobox')}</Alert>*/}
       {alert && <><Alert severity="error">{t('error_query_no_subcorpus_picked')}</Alert><br/></>}
       <div className="buttonBox">
         <Button variant="contained"
-                onClick={() => {
+                id="choose-text-button"
+                className={"button-query-hover button-box-open"}
+                onClick={(e) => {
                   setQueryVisible(true);
                   setExpanded(true);
-                }}>{t('query_choose_texts')}</Button>
-        <Button variant="contained"
-                onClick={() => setModalOpen(true)}
-                className="buttonSecondLeft">{t('query_own_texts')}</Button>
-        <Button variant="contained"
-                onClick={() => navigate('adding')}
-                className="buttonRight">{t('common_publish_your_text')}</Button>
+                  setFilterBoxClass();
+                }}>
+          <div className="button-text-query">
+            <ManageSearchIcon className="manage-search-icon"/>
+            <span className="manage-serach-incon-text">
+              {t('query_choose_texts')}
+            </span>
+          </div>
+        </Button>
+        <Box
+          id="menu-box-choose-text"
+          hidden={filterOpen}
+          className="menu-box-choose-text"
+        >
+          {!queryAnswer ?
+            <form action=""
+                  onSubmit={(e) => {
+                    e.preventDefault();
+                  }}
+                  id="vorm">
+              <div className="queryContainer">
+                <div>
+                  <b>{t('query_subcorpus')}</b>
+                  <br/><br/>
+                  <Checkbox
+                    style={{color: "#9C27B0"}}
+                    checked={corpusCheckboxStatus.all}
+                    onChange={alterAllCorpusCheckboxes}
+                  />
+                  <label>{t('query_subcorpus_all')}</label>
+                  <br/>
+                  <Checkbox
+                    style={{color: "#9C27B0"}}
+                    id="clWmOIrLa"
+                    checked={corpusCheckboxStatus.clWmOIrLa}
+                    onChange={alterCorpusCheckbox}
+                  />
+                  <Tooltip title={t('query_subcorpus_L2_proficiency_examinations_hover')}
+                           followCursor>
+                    <label className="corpustitle">
+                      {t('query_subcorpus_L2_proficiency_examinations')}
+                    </label>
+                  </Tooltip>
+                  <br/>
+                  <Checkbox
+                    style={{color: "#9C27B0"}}
+                    id="cFqPphvYi"
+                    checked={corpusCheckboxStatus.cFqPphvYi}
+                    onChange={alterCorpusCheckbox}
+                  />
+                  <Tooltip title={t('query_subcoprus_L2_olympiade_hover')}
+                           followCursor>
+                    <label className="corpustitle">
+                      {t('query_subcoprus_L2_olympiade')}
+                    </label>
+                  </Tooltip>
+                  <br/>
+                  <Checkbox
+                    style={{color: "#9C27B0"}}
+                    id="cFOoRQekA"
+                    checked={corpusCheckboxStatus.cFOoRQekA}
+                    onChange={alterCorpusCheckbox}
+                  />
+                  <Tooltip
+                    title={t('query_subcorpus_L2_estonian_hover')}
+                    followCursor>
+                    <label className="corpustitle">
+                      {t('query_subcorpus_L2_estonian')}
+                    </label>
+                  </Tooltip>
+                  <br/>
+                  <Checkbox
+                    style={{color: "#9C27B0"}}
+                    id="cYDRkpymb"
+                    checked={corpusCheckboxStatus.cYDRkpymb}
+                    onChange={alterCorpusCheckbox}
+                  />
+                  <Tooltip title={t('query_subcorpus_L1_estonian_hover')}
+                           followCursor>
+                    <label className="corpustitle">
+                      {t('query_subcorpus_L1_estonian')}
+                    </label>
+                  </Tooltip>
+                  <br/>
+                  <Checkbox
+                    style={{color: "#9C27B0"}}
+                    id="cgSRJPKTr"
+                    checked={corpusCheckboxStatus.cgSRJPKTr}
+                    onChange={alterCorpusCheckbox}
+                  />
+                  <Tooltip title={t('query_subcorpus_L1_russian_hover')}
+                           followCursor>
+                    <label className="corpustitle">
+                      {t('query_subcorpus_L1_russian')}
+                    </label>
+                  </Tooltip>
+                  <br/>
+                  <Checkbox
+                    style={{color: "#9C27B0"}}
+                    id="cZjHWUPtD"
+                    checked={corpusCheckboxStatus.cZjHWUPtD}
+                    onChange={alterCorpusCheckbox}
+                  />
+                  <Tooltip
+                    title={t('query_subcorpus_L3_russian_hover')}
+                    followCursor>
+                    <label className="corpustitle">
+                      {t('query_subcorpus_L3_russian')}
+                    </label>
+                  </Tooltip>
+                  <br/>
+                  <Checkbox
+                    style={{color: "#9C27B0"}}
+                    id="cwUSEqQLt"
+                    checked={corpusCheckboxStatus.cwUSEqQLt}
+                    onChange={alterCorpusCheckbox}
+                  />
+                  <Tooltip
+                    title={t('query_subcorpus_academic_estonian_hover')}
+                    followCursor>
+                    <label className="corpustitle">
+                      {t('query_subcorpus_academic_estonian')}
+                    </label>
+                  </Tooltip>
+                  <br/>
+                </div>
+                <div>
+                  <b>{t('common_text_data')}</b>
+                  <br/><br/>
+                  <FormControl className={classes.formControl}
+                               size="small">
+                    <InputLabel id="types-label">{t('query_text_data_type')}</InputLabel>
+                    <Select
+                      labelId="types-label"
+                      label={t('query_text_data_type')}
+                      multiple
+                      value={textTypes}
+                      name="types"
+                      renderValue={(textType) => textType.length > 1 ? t('query_text_data_type_selected_plural', {amount: textType.length}) : t('query_text_data_type_selected')}
+                      disabled={getSelectedCorpusList().length === 0}
+                      MenuProps={MenuProps}
+                    >
+                      {Object.keys(textTypesOptions).map((corpus) => (
+                        corpusCheckboxStatus[corpus] && Object.keys(textTypesOptions[corpus]).map((textType) => (
+                          typeof textTypesOptions[corpus][textType] === 'string'
+                            ? <MenuItem key={textType}
+                                        id={textType}
+                                        onClick={(e) => alterTextTypeHierarchyDropdown(e, true, null)}
+                                        value={textType}>
+                              <ListItemIcon>
+                                <Checkbox id={textType}
+                                          style={{color: "#9C27B0"}}
+                                          checked={textTypes.indexOf(textType) > -1}/>
+                              </ListItemIcon>
+                              <ListItemText id={textType}
+                                            primary={t(textTypesOptions[corpus][textType])}/>
+                            </MenuItem>
+                            : <span key={`${textType}_span`}>
+                            <MenuItem key={textType}
+                                      id={textType}
+                                      onClick={(e) => alterTextTypeHierarchyDropdown(e, false, corpus)}
+                                      value={textType}>
+                              <ListItemIcon>
+                                <Checkbox id={textType}
+                                          style={{color: "#9C27B0"}}
+                                          checked={checkTextTypeHierarchyCheckboxStatus(textType, corpus)}/>
+                              </ListItemIcon>
+                              <ListItemText id={textType}
+                                            primary={t(textType)}/>
+                            </MenuItem>
+                              {Object.keys(textTypesOptions[corpus][textType]).map((specificTextType) => (
+                                <MenuItem key={specificTextType}
+                                          id={specificTextType}
+                                          onClick={(e) => alterTextTypeHierarchyDropdown(e, true, null)}
+                                          value={specificTextType}
+                                          sx={{paddingLeft: '2rem'}}>
+                                  <ListItemIcon>
+                                    <Checkbox id={specificTextType}
+                                              style={{color: "#9C27B0"}}
+                                              checked={textTypes.indexOf(specificTextType) > -1}/>
+                                  </ListItemIcon>
+                                  <ListItemText id={specificTextType}
+                                                primary={t(textTypesOptions[corpus][textType][specificTextType])}/>
+                                </MenuItem>
+                              ))}
+                          </span>
+                        ))
+                      ))}
+                    </Select>
+                  </FormControl>
+                  <br/><br/>
+                  <FormControl size="small">
+                    <InputLabel id="language-label">{t('query_text_data_language')}</InputLabel>
+                    <Select
+                      sx={{minWidth: selectWidth}}
+                      labelId="language-label"
+                      name="language"
+                      value={singlePropertyData.language}
+                      label={t('query_text_data_language')}
+                      onClick={(e) => alterSinglePropertyData(e, 'language')}
+                    >
+                      {Object.keys(textLanguageOptions).map((lang) => (
+                        <MenuItem key={lang} value={lang}>{t(textLanguageOptions[lang])}</MenuItem>
+                      ))}
+                    </Select>
+                  </FormControl>
+                  <br/><br/>
+                  {checkIfOnlySpecificCorpusIsChecked('cwUSEqQLt')
+                    ? <>
+                      <FormControl size="small">
+                        <InputLabel id="domain-label">{t('common_text_data_field_of_research')}</InputLabel>
+                        <Select
+                          sx={{minWidth: selectWidth}}
+                          labelId="domain-label"
+                          name="domain"
+                          value={singlePropertyData.domain}
+                          label={t('common_text_data_field_of_research')}
+                          onClick={(e) => alterSinglePropertyData(e, 'domain')}
+                        >
+                          {Object.keys(domainOptions).map((domain) => (
+                            <MenuItem key={domain} value={domain}>{t(domainOptions[domain])}</MenuItem>
+                          ))}
+                        </Select>
+                      </FormControl>
+                      <br/><br/>
+                      <FormControl
+                        className={classes.formControl}
+                        size="small"
+                      >
+                        <InputLabel id="usedMultiMaterials-label">
+                          {t('query_text_data_used_study_or_supporting_materials')}
+                        </InputLabel>
+                        <Select
+                          labelId="usedMultiMaterials-label"
+                          label={t('query_text_data_used_study_or_supporting_materials')}
+                          multiple
+                          value={usedMultiMaterials}
+                          name="usedMultiMaterials"
+                          renderValue={(material) => material.length > 1 ? t('query_text_data_material_selected_plural', {amount: material.length}) : t('query_text_data_material_selected')}
+                          MenuProps={MenuProps}
+                        >
+                          {Object.keys(usedMaterialsMultiOptions).map((material) => (
+                            typeof usedMaterialsMultiOptions[material] === 'string'
+                              ? <MenuItem key={material}
+                                          id={material}
+                                          onClick={(e) => alterUsedMaterialsHierarchyDropdown(e, true)}
+                                          value={material}>
+                                <ListItemIcon>
+                                  <Checkbox id={material}
+                                            style={{color: "#9C27B0"}}
+                                            checked={usedMultiMaterials.indexOf(material) > -1}/>
+                                </ListItemIcon>
+                                <ListItemText id={material}
+                                              primary={t(usedMaterialsMultiOptions[material])}/>
+                              </MenuItem>
+                              : <span key={`${material}_span`}>
+                                <MenuItem key={material}
+                                          id={material}
+                                          onClick={(e) => alterUsedMaterialsHierarchyDropdown(e, false)}
+                                          value={material}>
+                                  <ListItemIcon>
+                                    <Checkbox id={material}
+                                              style={{color: "#9C27B0"}}
+                                              checked={checkUsedMaterialsHierarchyCheckboxStatus(material)}/>
+                                  </ListItemIcon>
+                                  <ListItemText id={material}
+                                                primary={t(material)}/>
+                                </MenuItem>
+                                {Object.keys(usedMaterialsMultiOptions[material]).map((subMaterial) => (
+                                  <MenuItem key={subMaterial}
+                                            id={subMaterial}
+                                            onClick={(e) => alterUsedMaterialsHierarchyDropdown(e, true)}
+                                            value={subMaterial}
+                                            sx={{paddingLeft: '2rem'}}>
+                                    <ListItemIcon>
+                                      <Checkbox id={subMaterial}
+                                                style={{color: "#9C27B0"}}
+                                                checked={usedMultiMaterials.indexOf(subMaterial) > -1}/>
+                                    </ListItemIcon>
+                                    <ListItemText id={subMaterial}
+                                                  primary={t(usedMaterialsMultiOptions[material][subMaterial])}/>
+                                  </MenuItem>
+                                ))}
+                              </span>
+                          ))}
+                        </Select>
+                      </FormControl>
+                    </>
+                    : <>
+                      <FormControl size="small">
+                        <InputLabel id="level-label">{t('query_text_data_level')}</InputLabel>
+                        <Select
+                          sx={{minWidth: selectWidth}}
+                          labelId="level-label"
+                          name="level"
+                          value={singlePropertyData.level}
+                          label={t('query_text_data_level')}
+                          onClick={(e) => alterSinglePropertyData(e, 'level')}
+                        >
+                          {textLevelOptions.map((level) => (
+                            <MenuItem key={level} value={level}>{level}</MenuItem>
+                          ))}
+                        </Select>
+                      </FormControl>
+                      <br/><br/>
+                      <FormControl size="small">
+                        <InputLabel
+                          id="usedMaterials-label">{t('query_text_data_used_supporting_materials')}</InputLabel>
+                        <Select
+                          sx={{minWidth: selectWidth}}
+                          labelId="usedMaterials-label"
+                          name="usedMaterials"
+                          value={singlePropertyData.usedMaterials}
+                          label={t('query_text_data_used_supporting_materials')}
+                          onClick={(e) => alterSinglePropertyData(e, 'usedMaterials')}
+                        >
+                          {Object.keys(usedMaterialsOptions).map((material) => (
+                            <MenuItem key={material} value={material}>{t(usedMaterialsOptions[material])}</MenuItem>
+                          ))}
+                        </Select>
+                      </FormControl>
+                    </>
+                  }
+                  <br/><br/>
+                  <FormControl className={classes.formControl}
+                               size="small">
+                    <InputLabel id="addedYears-label">{t('query_text_data_year_of_publication')}</InputLabel>
+                    <Select
+                      labelId="addedYears-label"
+                      label={t('query_text_data_year_of_publication')}
+                      multiple
+                      value={addedYears}
+                      name="addedYears"
+                      onChange={(e) => setAddedYears(e.target.value)}
+                      renderValue={(addedYear) => addedYear.join(', ')}
+                      MenuProps={MenuProps}
+                    >
+                      {addedYearOptions.map((year) => (
+                        <MenuItem key={year}
+                                  value={t(year)}>
+                          <ListItemIcon>
+                            <Checkbox style={{color: "#9C27B0"}} checked={addedYears.indexOf(t(year)) > -1}/>
+                          </ListItemIcon>
+                          <ListItemText primary={year}/>
+                        </MenuItem>
+                      ))}
+                    </Select>
+                  </FormControl>
+                  <br/><br/>
+                  <FormControl className={classes.formControl}
+                               size="small">
+                    <InputLabel id="characters-label">{t('query_text_data_characters')}</InputLabel>
+                    <Select
+                      labelId="characters-label"
+                      label={t('query_text_data_characters')}
+                      multiple
+                      value={characters}
+                      name="characters"
+                      onChange={(e) => setCharacters(e.target.value)}
+                      renderValue={(character) => character.join(', ')}
+                      MenuProps={MenuProps}
+                    >
+                      {charactersOptions.map((item) => (
+                        <MenuItem key={item}
+                                  value={t(item)}>
+                          <ListItemIcon>
+                            <Checkbox style={{color: "#9C27B0"}} checked={characters.indexOf(t(item)) > -1}/>
+                          </ListItemIcon>
+                          <ListItemText primary={t(item)}/>
+                        </MenuItem>
+                      ))}
+                    </Select>
+                  </FormControl>
+                  <br/><br/>
+                  <FormControl className={classes.formControl}
+                               size="small">
+                    <InputLabel id="words-label">{t('common_words')}</InputLabel>
+                    <Select
+                      labelId="words-label"
+                      label={t('common_words')}
+                      multiple
+                      value={words}
+                      name="words"
+                      onChange={(e) => setWords(e.target.value)}
+                      renderValue={(word) => word.join(', ')}
+                      MenuProps={MenuProps}
+                    >
+                      {wordsOptions.map((item) => (
+                        <MenuItem key={item}
+                                  value={t(item)}>
+                          <ListItemIcon>
+                            <Checkbox style={{color: "#9C27B0"}} checked={words.indexOf(t(item)) > -1}/>
+                          </ListItemIcon>
+                          <ListItemText primary={t(item)}/>
+                        </MenuItem>
+                      ))}
+                    </Select>
+                  </FormControl>
+                  <br/><br/>
+                  <FormControl className={classes.formControl}
+                               size="small">
+                    <InputLabel id="sentences-label">{t('common_sentences')}</InputLabel>
+                    <Select
+                      labelId="sentences-label"
+                      label={t('common_sentences')}
+                      multiple
+                      value={sentences}
+                      name="sentences"
+                      onChange={(e) => setSentences(e.target.value)}
+                      renderValue={(sentence) => sentence.join(', ')}
+                      MenuProps={MenuProps}
+                    >
+                      {sentencesOptions.map((item) => (
+                        <MenuItem key={item}
+                                  value={t(item)}>
+                          <ListItemIcon>
+                            <Checkbox style={{color: "#9C27B0"}} checked={sentences.indexOf(t(item)) > -1}/>
+                          </ListItemIcon>
+                          <ListItemText primary={t(item)}/>
+                        </MenuItem>
+                      ))}
+                    </Select>
+                  </FormControl>
+                </div>
+                <div>
+                  <b>{t('common_author_data')}</b>
+                  <br/><br/>
+                  <FormControl size="small">
+                    <InputLabel id="age-label">{t('query_author_data_age')}</InputLabel>
+                    <Select
+                      sx={{minWidth: selectWidth}}
+                      labelId="age-label"
+                      name="age"
+                      value={singlePropertyData.age}
+                      label={t('query_author_data_age')}
+                      onClick={(e) => alterSinglePropertyData(e, 'age')}
+                    >
+                      {Object.keys(ageOptions).map((age) => (
+                        <MenuItem key={age} value={age}>{t(ageOptions[age])}</MenuItem>
+                      ))}
+                    </Select>
+                  </FormControl>
+                  <br/><br/>
+                  <FormControl size="small">
+                    <InputLabel id="gender-label">{t('query_author_data_gender')}</InputLabel>
+                    <Select
+                      sx={{minWidth: selectWidth}}
+                      labelId="gender-label"
+                      name="gender"
+                      value={singlePropertyData.gender}
+                      label={t('query_author_data_gender')}
+                      onClick={(e) => alterSinglePropertyData(e, 'gender')}
+                    >
+                      {Object.keys(genderOptions).map((gender) => (
+                        <MenuItem key={gender} value={gender}>{t(genderOptions[gender])}</MenuItem>
+                      ))}
+                    </Select>
+                  </FormControl>
+                  <br/><br/>
+                  {checkIfOnlySpecificCorpusIsChecked('cwUSEqQLt')
+                    ? <>
+                      <FormControl size="small">
+                        <InputLabel id="studyLevel-label">{t('query_author_data_level_of_study')}</InputLabel>
+                        <Select
+                          sx={{minWidth: selectWidth}}
+                          labelId="studyLevel-label"
+                          name="studyLevel"
+                          value={singlePropertyData.studyLevel}
+                          label={t('query_author_data_level_of_study')}
+                          onClick={(e) => alterSinglePropertyData(e, 'studyLevel')}
+                        >
+                          {Object.keys(studyLevelOptions).map((level) => (
+                            <MenuItem key={level} value={level}>{t(studyLevelOptions[level])}</MenuItem>
+                          ))}
+                        </Select>
+                      </FormControl>
+                      <br/><br/>
+                      <FormControl size="small">
+                        <InputLabel id="degree-label">{t('query_author_data_degree')}</InputLabel>
+                        <Select
+                          sx={{minWidth: selectWidth}}
+                          labelId="degree-label"
+                          name="degree"
+                          value={singlePropertyData.degree}
+                          label={t('query_author_data_degree')}
+                          onClick={(e) => alterSinglePropertyData(e, 'degree')}
+                        >
+                          {Object.keys(degreeOptions).map((degree) => (
+                            <MenuItem key={degree} value={degree}>{t(degreeOptions[degree])}</MenuItem>
+                          ))}
+                        </Select>
+                      </FormControl>
+                    </>
+                    : <FormControl size="small">
+                      <InputLabel id="education-label">{t('query_author_data_education')}</InputLabel>
+                      <Select
+                        sx={{minWidth: selectWidth}}
+                        labelId="education-label"
+                        name="education"
+                        value={singlePropertyData.education}
+                        label={t('query_author_data_education')}
+                        onClick={(e) => alterSinglePropertyData(e, 'education')}
+                      >
+                        {Object.keys(educationOptions).map((education) => (
+                          <MenuItem key={education} value={education}>{t(educationOptions[education])}</MenuItem>
+                        ))}
+                      </Select>
+                    </FormControl>}
+                  <br/><br/>
+                  {checkIfOnlySpecificCorpusIsChecked('clWmOIrLa')
+                    ? <FormControl size="small">
+                      <InputLabel id="nationality-label">{t('query_author_data_nationality')}</InputLabel>
+                      <Select
+                        sx={{minWidth: selectWidth}}
+                        labelId="nationality-label"
+                        name="nationality"
+                        value={singlePropertyData.nationality}
+                        label={t('query_author_data_nationality')}
+                        onClick={(e) => alterSinglePropertyData(e, 'nationality')}
+                      >
+                        {Object.keys(nationalityOptions).map((nationality) => (
+                          <MenuItem key={nationality}
+                                    value={nationality}>{t(nationalityOptions[nationality])}</MenuItem>
+                        ))}
+                      </Select>
+                    </FormControl>
+                    : <FormControl size="small">
+                      <InputLabel id="nativeLang-label">{t('query_author_data_native_language')}</InputLabel>
+                      <Select
+                        sx={{minWidth: selectWidth}}
+                        labelId="nativeLang-label"
+                        name="nativeLang"
+                        value={singlePropertyData.nativeLang}
+                        label={t('query_author_data_native_language')}
+                        onClick={(e) => alterSinglePropertyData(e, 'nativeLang')}
+                      >
+                        {Object.keys(languageOptions).map((lang) => (
+                          <MenuItem key={lang} value={lang}>{t(languageOptions[lang])}</MenuItem>
+                        ))}
+                      </Select>
+                    </FormControl>
+                  }
+                  <br/><br/>
+                  {checkIfOnlySpecificCorpusIsChecked('cwUSEqQLt')
+                    ? <>
+                      <FormControl size="small">
+                        <InputLabel id="otherLang-label">{t('query_author_data_other_languages')}</InputLabel>
+                        <Select
+                          sx={{minWidth: selectWidth}}
+                          labelId="otherLang-label"
+                          name="otherLang"
+                          value={singlePropertyData.otherLang}
+                          label={t('query_author_data_other_languages')}
+                          onClick={(e) => alterSinglePropertyData(e, 'otherLang')}
+                        >
+                          {Object.keys(languageOptions).map((lang) => (
+                            <MenuItem key={lang} value={lang}>{t(languageOptions[lang])}</MenuItem>
+                          ))}
+                        </Select>
+                      </FormControl>
+                      <br/><br/>
+                    </>
+                    : <></>}
+                  <FormControl size="small">
+                    <InputLabel id="country-label">{t('query_author_data_country')}</InputLabel>
+                    <Select
+                      sx={{minWidth: selectWidth}}
+                      labelId="country-label"
+                      name="country"
+                      value={singlePropertyData.country}
+                      label={t('query_author_data_country')}
+                      onClick={(e) => alterSinglePropertyData(e, 'country')}
+                    >
+                      {Object.keys(countryOptions).map((country) => (
+                        <MenuItem key={country} value={country}>{t(countryOptions[country])}</MenuItem>
+                      ))}
+                    </Select>
+                  </FormControl>
+                </div>
+              </div>
+              <br/><br/>
+              <Button onClick={submitted}
+                      sx={{bgcolor: "#9C27B0", fontWeight: "bold", borderRadius: "15px"}}
+                      variant="contained">{t('send_request_button')}</Button>
+            </form> : <span>
+              {noResultsError &&
+                <div><br/><Alert severity="error">{t('query_results_no_texts_found')}</Alert><br/></div>}
+              <QueryResults key={resultsKey}
+                            data={results}
+                            setFilterBoxClass={() => setFilterBoxClass()}/>
+        </span>}
+        </Box>
+        <Button id="choose-input-button" variant="contained"
+                className={"button-query-hover button-box-open"}
+                onClick={() => setInputHidden()}
+        >
+          <div className="button-text-query"><ReadMoreIcon
+            style={{fontSize: "40px", marginLeft: "10px", marginRight: "15px"}}/><span
+            style={{fontSize: "15px", fontWeight: "bold"}}>{t('query_own_texts')}</span></div>
+        </Button>
+        <Box hidden={inputOpen} className="query-own-texts-modal">
+          <div>
+            {t('textupload_primary_modal_title')}
+          </div>
+          <br/>
+          <div>
+            <TextUpload sendTextFromFile={sendTextFromFile}/>
+            <textarea
+              spellCheck="false"
+              className="query-textinput"
+              value={textInputValue}
+              onChange={(e) => setTextInputValue(e.target.value)}
+            ></textarea>
+            <Button variant="contained"
+                    sx={{bgcolor: "#9C27B0", fontWeight: "bold", borderRadius: "15px"}}
+                    disabled={textInputValue === ''}
+                    onClick={() => {
+                      handleSubmitOwnTexts();
+                      setInputHidden()
+                    }}>
+              {t('textupload_primary_modal_save')}
+            </Button>
+          </div>
+        </Box>
+        {/*<Button variant="contained"
+                 onClick={() => navigate('adding')}
+                 className="buttonRight">{t('common_publish_your_text')}</Button>*/}
       </div>
-      {(corpusTextsSelected > 0 || ownTextsSelected) && <>
+      {/*{(corpusTextsSelected > 0 || ownTextsSelected) && <>
         <br/>
         {t('query_results_saved_for_analysis')}
         {corpusTextsSelected > 0 &&
@@ -429,615 +1082,7 @@ export default function Query() {
                 variant="outlined"
                 onDelete={() => handleChipDelete('OWN_TEXTS')}/>
         }
-      </>}
-      {queryVisible && <span>
-        <Accordion sx={AccordionStyle}
-                   className="queryAccordion"
-                   expanded={expanded}
-                   onChange={changeAccordion}>
-        <AccordionSummary
-          expandIcon={<ExpandMoreIcon/>}
-          id="filters-header"
-        >
-          <Typography>
-            {t('query_filters')}
-          </Typography>
-        </AccordionSummary>
-        <AccordionDetails>
-          <form action=""
-                onSubmit={(e) => {
-                  e.preventDefault();
-                }}
-                id="vorm">
-            <div className="queryContainer">
-              <div>
-                <b>{t('query_subcorpus')}</b>
-                <br/><br/>
-                <Checkbox
-                  checked={corpusCheckboxStatus.all}
-                  onChange={alterAllCorpusCheckboxes}
-                />
-                <label>{t('query_subcorpus_all')}</label>
-                <br/>
-                <Checkbox
-                  id="clWmOIrLa"
-                  checked={corpusCheckboxStatus.clWmOIrLa}
-                  onChange={alterCorpusCheckbox}
-                />
-                <Tooltip title={t('query_subcorpus_L2_proficiency_examinations_hover')}
-                         followCursor>
-                  <label className="corpustitle">
-                    {t('query_subcorpus_L2_proficiency_examinations')}
-                  </label>
-                </Tooltip>
-                <br/>
-                <Checkbox
-                  id="cFqPphvYi"
-                  checked={corpusCheckboxStatus.cFqPphvYi}
-                  onChange={alterCorpusCheckbox}
-                />
-                <Tooltip title={t('query_subcoprus_L2_olympiade_hover')}
-                         followCursor>
-                  <label className="corpustitle">
-                    {t('query_subcoprus_L2_olympiade')}
-                  </label>
-                </Tooltip>
-                <br/>
-                <Checkbox
-                  id="cFOoRQekA"
-                  checked={corpusCheckboxStatus.cFOoRQekA}
-                  onChange={alterCorpusCheckbox}
-                />
-                <Tooltip
-                  title={t('query_subcorpus_L2_estonian_hover')}
-                  followCursor>
-                  <label className="corpustitle">
-                    {t('query_subcorpus_L2_estonian')}
-                  </label>
-                </Tooltip>
-                <br/>
-                <Checkbox
-                  id="cYDRkpymb"
-                  checked={corpusCheckboxStatus.cYDRkpymb}
-                  onChange={alterCorpusCheckbox}
-                />
-                <Tooltip title={t('query_subcorpus_L1_estonian_hover')}
-                         followCursor>
-                  <label className="corpustitle">
-                    {t('query_subcorpus_L1_estonian')}
-                  </label>
-                </Tooltip>
-                <br/>
-                <Checkbox
-                  id="cgSRJPKTr"
-                  checked={corpusCheckboxStatus.cgSRJPKTr}
-                  onChange={alterCorpusCheckbox}
-                />
-                <Tooltip title={t('query_subcorpus_L1_russian_hover')}
-                         followCursor>
-                  <label className="corpustitle">
-                    {t('query_subcorpus_L1_russian')}
-                  </label>
-                </Tooltip>
-                <br/>
-                <Checkbox
-                  id="cZjHWUPtD"
-                  checked={corpusCheckboxStatus.cZjHWUPtD}
-                  onChange={alterCorpusCheckbox}
-                />
-                <Tooltip
-                  title={t('query_subcorpus_L3_russian_hover')}
-                  followCursor>
-                  <label className="corpustitle">
-                    {t('query_subcorpus_L3_russian')}
-                  </label>
-                </Tooltip>
-                <br/>
-                <Checkbox
-                  id="cwUSEqQLt"
-                  checked={corpusCheckboxStatus.cwUSEqQLt}
-                  onChange={alterCorpusCheckbox}
-                />
-                <Tooltip
-                  title={t('query_subcorpus_academic_estonian_hover')}
-                  followCursor>
-                  <label className="corpustitle">
-                    {t('query_subcorpus_academic_estonian')}
-                  </label>
-                </Tooltip>
-                <br/>
-              </div>
-              <div>
-                <b>{t('common_text_data')}</b>
-                <br/><br/>
-                <FormControl className={classes.formControl}
-                             size="small">
-                  <InputLabel id="types-label">{t('query_text_data_type')}</InputLabel>
-                  <Select
-                    labelId="types-label"
-                    label={t('query_text_data_type')}
-                    multiple
-                    value={textTypes}
-                    name="types"
-                    renderValue={(textType) => textType.length > 1 ? t('query_text_data_type_selected_plural', {amount: textType.length}) : t('query_text_data_type_selected')}
-                    disabled={getSelectedCorpusList().length === 0}
-                    MenuProps={MenuProps}
-                  >
-                    {Object.keys(textTypesOptions).map((corpus) => (
-                      corpusCheckboxStatus[corpus] && Object.keys(textTypesOptions[corpus]).map((textType) => (
-                        typeof textTypesOptions[corpus][textType] === 'string'
-                          ? <MenuItem key={textType}
-                                      id={textType}
-                                      onClick={(e) => alterTextTypeHierarchyDropdown(e, true, null)}
-                                      value={textType}>
-                            <ListItemIcon>
-                              <Checkbox id={textType}
-                                        checked={textTypes.indexOf(textType) > -1}/>
-                            </ListItemIcon>
-                            <ListItemText id={textType}
-                                          primary={t(textTypesOptions[corpus][textType])}/>
-                          </MenuItem>
-                          : <span key={`${textType}_span`}>
-                            <MenuItem key={textType}
-                                      id={textType}
-                                      onClick={(e) => alterTextTypeHierarchyDropdown(e, false, corpus)}
-                                      value={textType}>
-                              <ListItemIcon>
-                                <Checkbox id={textType}
-                                          checked={checkTextTypeHierarchyCheckboxStatus(textType, corpus)}/>
-                              </ListItemIcon>
-                              <ListItemText id={textType}
-                                            primary={t(textType)}/>
-                            </MenuItem>
-                            {Object.keys(textTypesOptions[corpus][textType]).map((specificTextType) => (
-                              <MenuItem key={specificTextType}
-                                        id={specificTextType}
-                                        onClick={(e) => alterTextTypeHierarchyDropdown(e, true, null)}
-                                        value={specificTextType}
-                                        sx={{paddingLeft: '2rem'}}>
-                                <ListItemIcon>
-                                  <Checkbox id={specificTextType}
-                                            checked={textTypes.indexOf(specificTextType) > -1}/>
-                                </ListItemIcon>
-                                <ListItemText id={specificTextType}
-                                              primary={t(textTypesOptions[corpus][textType][specificTextType])}/>
-                              </MenuItem>
-                            ))}
-                          </span>
-                      ))
-                    ))}
-                  </Select>
-                </FormControl>
-                <br/><br/>
-                <FormControl size="small">
-                  <InputLabel id="language-label">{t('query_text_data_language')}</InputLabel>
-                  <Select
-                    sx={{minWidth: selectWidth}}
-                    labelId="language-label"
-                    name="language"
-                    value={singlePropertyData.language}
-                    label={t('query_text_data_language')}
-                    onClick={(e) => alterSinglePropertyData(e, 'language')}
-                  >
-                    {Object.keys(textLanguageOptions).map((lang) => (
-                      <MenuItem key={lang} value={lang}>{t(textLanguageOptions[lang])}</MenuItem>
-                    ))}
-                  </Select>
-                </FormControl>
-                <br/><br/>
-                {checkIfOnlySpecificCorpusIsChecked('cwUSEqQLt')
-                  ? <>
-                    <FormControl size="small">
-                      <InputLabel id="domain-label">{t('common_text_data_field_of_research')}</InputLabel>
-                      <Select
-                        sx={{minWidth: selectWidth}}
-                        labelId="domain-label"
-                        name="domain"
-                        value={singlePropertyData.domain}
-                        label={t('common_text_data_field_of_research')}
-                        onClick={(e) => alterSinglePropertyData(e, 'domain')}
-                      >
-                        {Object.keys(domainOptions).map((domain) => (
-                          <MenuItem key={domain} value={domain}>{t(domainOptions[domain])}</MenuItem>
-                        ))}
-                      </Select>
-                    </FormControl>
-                    <br/><br/>
-                    <FormControl className={classes.formControl}
-                                 size="small">
-                      <InputLabel
-                        id="usedMultiMaterials-label">{t('query_text_data_used_study_or_supporting_materials')}</InputLabel>
-                      <Select
-                        labelId="usedMultiMaterials-label"
-                        label={t('query_text_data_used_study_or_supporting_materials')}
-                        multiple
-                        value={usedMultiMaterials}
-                        name="usedMultiMaterials"
-                        renderValue={(material) => material.length > 1 ? t('query_text_data_material_selected_plural', {amount: material.length}) : t('query_text_data_material_selected')}
-                        MenuProps={MenuProps}
-                      >
-                        {Object.keys(usedMaterialsMultiOptions).map((material) => (
-                          typeof usedMaterialsMultiOptions[material] === 'string'
-                            ? <MenuItem key={material}
-                                        id={material}
-                                        onClick={(e) => alterUsedMaterialsHierarchyDropdown(e, true)}
-                                        value={material}>
-                              <ListItemIcon>
-                                <Checkbox id={material}
-                                          checked={usedMultiMaterials.indexOf(material) > -1}/>
-                              </ListItemIcon>
-                              <ListItemText id={material}
-                                            primary={t(usedMaterialsMultiOptions[material])}/>
-                            </MenuItem>
-                            : <span key={`${material}_span`}>
-                                <MenuItem key={material}
-                                          id={material}
-                                          onClick={(e) => alterUsedMaterialsHierarchyDropdown(e, false)}
-                                          value={material}>
-                                  <ListItemIcon>
-                                    <Checkbox id={material}
-                                              checked={checkUsedMaterialsHierarchyCheckboxStatus(material)}/>
-                                  </ListItemIcon>
-                                  <ListItemText id={material}
-                                                primary={t(material)}/>
-                                </MenuItem>
-                              {Object.keys(usedMaterialsMultiOptions[material]).map((subMaterial) => (
-                                <MenuItem key={subMaterial}
-                                          id={subMaterial}
-                                          onClick={(e) => alterUsedMaterialsHierarchyDropdown(e, true)}
-                                          value={subMaterial}
-                                          sx={{paddingLeft: '2rem'}}>
-                                  <ListItemIcon>
-                                    <Checkbox id={subMaterial}
-                                              checked={usedMultiMaterials.indexOf(subMaterial) > -1}/>
-                                  </ListItemIcon>
-                                  <ListItemText id={subMaterial}
-                                                primary={t(usedMaterialsMultiOptions[material][subMaterial])}/>
-                                </MenuItem>
-                              ))}
-                              </span>
-                        ))}
-                      </Select>
-                    </FormControl>
-                  </>
-                  : <>
-                    <FormControl size="small">
-                      <InputLabel id="level-label">{t('query_text_data_level')}</InputLabel>
-                      <Select
-                        sx={{minWidth: selectWidth}}
-                        labelId="level-label"
-                        name="level"
-                        value={singlePropertyData.level}
-                        label={t('query_text_data_level')}
-                        onClick={(e) => alterSinglePropertyData(e, 'level')}
-                      >
-                        {textLevelOptions.map((level) => (
-                          <MenuItem key={level} value={level}>{level}</MenuItem>
-                        ))}
-                      </Select>
-                    </FormControl>
-                    <br/><br/>
-                    <FormControl size="small">
-                      <InputLabel id="usedMaterials-label">{t('query_text_data_used_supporting_materials')}</InputLabel>
-                      <Select
-                        sx={{minWidth: selectWidth}}
-                        labelId="usedMaterials-label"
-                        name="usedMaterials"
-                        value={singlePropertyData.usedMaterials}
-                        label={t('query_text_data_used_supporting_materials')}
-                        onClick={(e) => alterSinglePropertyData(e, 'usedMaterials')}
-                      >
-                        {Object.keys(usedMaterialsOptions).map((material) => (
-                          <MenuItem key={material} value={material}>{t(usedMaterialsOptions[material])}</MenuItem>
-                        ))}
-                      </Select>
-                    </FormControl>
-                  </>
-                }
-                <br/><br/>
-                <FormControl className={classes.formControl}
-                             size="small">
-                  <InputLabel id="addedYears-label">{t('query_text_data_year_of_publication')}</InputLabel>
-                  <Select
-                    labelId="addedYears-label"
-                    label={t('query_text_data_year_of_publication')}
-                    multiple
-                    value={addedYears}
-                    name="addedYears"
-                    onChange={(e) => setAddedYears(e.target.value)}
-                    renderValue={(addedYear) => addedYear.join(', ')}
-                    MenuProps={MenuProps}
-                  >
-                    {addedYearOptions.map((year) => (
-                      <MenuItem key={year}
-                                value={t(year)}>
-                        <ListItemIcon>
-                          <Checkbox checked={addedYears.indexOf(t(year)) > -1}/>
-                        </ListItemIcon>
-                        <ListItemText primary={year}/>
-                      </MenuItem>
-                    ))}
-                  </Select>
-                </FormControl>
-                <br/><br/>
-                <FormControl className={classes.formControl}
-                             size="small">
-                  <InputLabel id="characters-label">{t('query_text_data_characters')}</InputLabel>
-                  <Select
-                    labelId="characters-label"
-                    label={t('query_text_data_characters')}
-                    multiple
-                    value={characters}
-                    name="characters"
-                    onChange={(e) => setCharacters(e.target.value)}
-                    renderValue={(character) => character.join(', ')}
-                    MenuProps={MenuProps}
-                  >
-                    {charactersOptions.map((item) => (
-                      <MenuItem key={item}
-                                value={t(item)}>
-                        <ListItemIcon>
-                          <Checkbox checked={characters.indexOf(t(item)) > -1}/>
-                        </ListItemIcon>
-                        <ListItemText primary={t(item)}/>
-                      </MenuItem>
-                    ))}
-                  </Select>
-                </FormControl>
-                <br/><br/>
-                <FormControl className={classes.formControl}
-                             size="small">
-                  <InputLabel id="words-label">{t('common_words')}</InputLabel>
-                  <Select
-                    labelId="words-label"
-                    label={t('common_words')}
-                    multiple
-                    value={words}
-                    name="words"
-                    onChange={(e) => setWords(e.target.value)}
-                    renderValue={(word) => word.join(', ')}
-                    MenuProps={MenuProps}
-                  >
-                    {wordsOptions.map((item) => (
-                      <MenuItem key={item}
-                                value={t(item)}>
-                        <ListItemIcon>
-                          <Checkbox checked={words.indexOf(t(item)) > -1}/>
-                        </ListItemIcon>
-                        <ListItemText primary={t(item)}/>
-                      </MenuItem>
-                    ))}
-                  </Select>
-                </FormControl>
-                <br/><br/>
-                <FormControl className={classes.formControl}
-                             size="small">
-                  <InputLabel id="sentences-label">{t('common_sentences')}</InputLabel>
-                  <Select
-                    labelId="sentences-label"
-                    label={t('common_sentences')}
-                    multiple
-                    value={sentences}
-                    name="sentences"
-                    onChange={(e) => setSentences(e.target.value)}
-                    renderValue={(sentence) => sentence.join(', ')}
-                    MenuProps={MenuProps}
-                  >
-                    {sentencesOptions.map((item) => (
-                      <MenuItem key={item}
-                                value={t(item)}>
-                        <ListItemIcon>
-                          <Checkbox checked={sentences.indexOf(t(item)) > -1}/>
-                        </ListItemIcon>
-                        <ListItemText primary={t(item)}/>
-                      </MenuItem>
-                    ))}
-                  </Select>
-                </FormControl>
-              </div>
-              <div>
-                <b>{t('common_author_data')}</b>
-                <br/><br/>
-                <FormControl size="small">
-                  <InputLabel id="age-label">{t('query_author_data_age')}</InputLabel>
-                  <Select
-                    sx={{minWidth: selectWidth}}
-                    labelId="age-label"
-                    name="age"
-                    value={singlePropertyData.age}
-                    label={t('query_author_data_age')}
-                    onClick={(e) => alterSinglePropertyData(e, 'age')}
-                  >
-                    {Object.keys(ageOptions).map((age) => (
-                      <MenuItem key={age} value={age}>{t(ageOptions[age])}</MenuItem>
-                    ))}
-                  </Select>
-                </FormControl>
-                <br/><br/>
-                <FormControl size="small">
-                  <InputLabel id="gender-label">{t('query_author_data_gender')}</InputLabel>
-                  <Select
-                    sx={{minWidth: selectWidth}}
-                    labelId="gender-label"
-                    name="gender"
-                    value={singlePropertyData.gender}
-                    label={t('query_author_data_gender')}
-                    onClick={(e) => alterSinglePropertyData(e, 'gender')}
-                  >
-                    {Object.keys(genderOptions).map((gender) => (
-                      <MenuItem key={gender} value={gender}>{t(genderOptions[gender])}</MenuItem>
-                    ))}
-                  </Select>
-                </FormControl>
-                <br/><br/>
-                {checkIfOnlySpecificCorpusIsChecked('cwUSEqQLt')
-                  ? <>
-                    <FormControl size="small">
-                      <InputLabel id="studyLevel-label">{t('query_author_data_level_of_study')}</InputLabel>
-                      <Select
-                        sx={{minWidth: selectWidth}}
-                        labelId="studyLevel-label"
-                        name="studyLevel"
-                        value={singlePropertyData.studyLevel}
-                        label={t('query_author_data_level_of_study')}
-                        onClick={(e) => alterSinglePropertyData(e, 'studyLevel')}
-                      >
-                        {Object.keys(studyLevelOptions).map((level) => (
-                          <MenuItem key={level} value={level}>{t(studyLevelOptions[level])}</MenuItem>
-                        ))}
-                      </Select>
-                    </FormControl>
-                    <br/><br/>
-                    <FormControl size="small">
-                      <InputLabel id="degree-label">{t('query_author_data_degree')}</InputLabel>
-                      <Select
-                        sx={{minWidth: selectWidth}}
-                        labelId="degree-label"
-                        name="degree"
-                        value={singlePropertyData.degree}
-                        label={t('query_author_data_degree')}
-                        onClick={(e) => alterSinglePropertyData(e, 'degree')}
-                      >
-                        {Object.keys(degreeOptions).map((degree) => (
-                          <MenuItem key={degree} value={degree}>{t(degreeOptions[degree])}</MenuItem>
-                        ))}
-                      </Select>
-                    </FormControl>
-                  </>
-                  : <FormControl size="small">
-                    <InputLabel id="education-label">{t('query_author_data_education')}</InputLabel>
-                    <Select
-                      sx={{minWidth: selectWidth}}
-                      labelId="education-label"
-                      name="education"
-                      value={singlePropertyData.education}
-                      label={t('query_author_data_education')}
-                      onClick={(e) => alterSinglePropertyData(e, 'education')}
-                    >
-                      {Object.keys(educationOptions).map((education) => (
-                        <MenuItem key={education} value={education}>{t(educationOptions[education])}</MenuItem>
-                      ))}
-                    </Select>
-                  </FormControl>}
-                <br/><br/>
-                {checkIfOnlySpecificCorpusIsChecked('clWmOIrLa')
-                  ? <FormControl size="small">
-                    <InputLabel id="nationality-label">{t('query_author_data_nationality')}</InputLabel>
-                    <Select
-                      sx={{minWidth: selectWidth}}
-                      labelId="nationality-label"
-                      name="nationality"
-                      value={singlePropertyData.nationality}
-                      label={t('query_author_data_nationality')}
-                      onClick={(e) => alterSinglePropertyData(e, 'nationality')}
-                    >
-                      {Object.keys(nationalityOptions).map((nationality) => (
-                        <MenuItem key={nationality} value={nationality}>{t(nationalityOptions[nationality])}</MenuItem>
-                      ))}
-                    </Select>
-                  </FormControl>
-                  : <FormControl size="small">
-                    <InputLabel id="nativeLang-label">{t('query_author_data_native_language')}</InputLabel>
-                    <Select
-                      sx={{minWidth: selectWidth}}
-                      labelId="nativeLang-label"
-                      name="nativeLang"
-                      value={singlePropertyData.nativeLang}
-                      label={t('query_author_data_native_language')}
-                      onClick={(e) => alterSinglePropertyData(e, 'nativeLang')}
-                    >
-                      {Object.keys(languageOptions).map((lang) => (
-                        <MenuItem key={lang} value={lang}>{t(languageOptions[lang])}</MenuItem>
-                      ))}
-                    </Select>
-                  </FormControl>
-                }
-                <br/><br/>
-                {checkIfOnlySpecificCorpusIsChecked('cwUSEqQLt')
-                  ? <>
-                    <FormControl size="small">
-                      <InputLabel id="otherLang-label">{t('query_author_data_other_languages')}</InputLabel>
-                      <Select
-                        sx={{minWidth: selectWidth}}
-                        labelId="otherLang-label"
-                        name="otherLang"
-                        value={singlePropertyData.otherLang}
-                        label={t('query_author_data_other_languages')}
-                        onClick={(e) => alterSinglePropertyData(e, 'otherLang')}
-                      >
-                        {Object.keys(languageOptions).map((lang) => (
-                          <MenuItem key={lang} value={lang}>{t(languageOptions[lang])}</MenuItem>
-                        ))}
-                      </Select>
-                    </FormControl>
-                    <br/><br/>
-                  </>
-                  : <></>}
-                <FormControl size="small">
-                  <InputLabel id="country-label">{t('query_author_data_country')}</InputLabel>
-                  <Select
-                    sx={{minWidth: selectWidth}}
-                    labelId="country-label"
-                    name="country"
-                    value={singlePropertyData.country}
-                    label={t('query_author_data_country')}
-                    onClick={(e) => alterSinglePropertyData(e, 'country')}
-                  >
-                    {Object.keys(countryOptions).map((country) => (
-                      <MenuItem key={country} value={country}>{t(countryOptions[country])}</MenuItem>
-                    ))}
-                  </Select>
-                </FormControl>
-              </div>
-            </div>
-            <br/><br/>
-            <Button onClick={submitted}
-                    variant="contained">{t('send_request_button')}</Button>
-          </form>
-        </AccordionDetails>
-      </Accordion>
-        {noResultsError && <div><br/><Alert severity="error">{t('query_results_no_texts_found')}</Alert><br/></div>}
-        <QueryResults key={resultsKey}
-                      data={results}/>
-        </span>
-      }
-      <Modal
-        open={modalOpen}
-        onClose={() => {
-          setModalOpen(false);
-        }}
-      >
-        <Box sx={modalStyle} className="query-own-texts-modal">
-          <div className="modal-head">
-            {t('textupload_primary_modal_title')}
-          </div>
-          <IconButton
-            aria-label="close"
-            onClick={() => {
-              setModalOpen(false);
-            }}
-            className="closeButton"
-          >
-            <CloseIcon/>
-          </IconButton>
-          <br/><br/>
-          <div>
-            <TextUpload sendTextFromFile={sendTextFromFile}/>
-            <textarea
-              spellCheck="false"
-              className="query-textinput"
-              value={textInputValue}
-              onChange={(e) => setTextInputValue(e.target.value)}
-            ></textarea>
-            <Button variant="contained"
-                    style={{marginTop: '2.5em'}}
-                    disabled={textInputValue === ''}
-                    onClick={() => handleSubmitOwnTexts()}>
-              {t('textupload_primary_modal_save')}
-            </Button>
-          </div>
-        </Box>
-      </Modal>
+      </>}*/}
     </div>
   );
 }
