+++ ui/src/elle/pages/styles/QueryResults.css
@@ -1,54 +1,61 @@
 .modal-head {
-  font-size: 1.5em;
+    font-size: 1.5em;
 }
 
-.resultTable {
-  margin-right: auto;
-  margin-left: auto;
-  width: 100%;
+.result-table {
+    margin-right: auto;
+    margin-left: auto;
+    width: 100%;
 }
 
-.tableRow {
-  background-color: white;
+.query-table-row {
+    background-color: white;
 }
 
-.tableRow:nth-child(even) {
-  background-color: #EBECF0;
+.query-table-row:nth-child(even) {
+    background-color: #EBECF0;
 }
 
-.checkboxRow {
-  width: 4em;
+.checkbox-row {
+    width: 4em;
 }
 
-.clickableRow {
-  cursor: pointer;
+.clickable-row {
+    cursor: pointer;
 }
 
-.closeButton {
-  position: absolute !important;
-  top: 0.5em;
-  right: 0.5em;
+.close-button {
+    position: absolute !important;
+    top: 0.5em;
+    right: 0.5em;
 }
 
-.selectAllButton {
-  margin-top: 1em !important;
-  margin-bottom: 1em !important;
+.select-all-button {
+    margin-top: 1em !important;
+    margin-bottom: 1em !important;
+    border-color: #9C27B0 !important;
+    color: #9C27B0 !important;
+    font-weight: bold !important;
+    border-radius: 15px !important;
 }
 
-.metainfoSubtitle {
-  font-weight: bold;
-  text-decoration: underline;
-  padding-bottom: 0.75em;
+.metainfo-subtitle {
+    font-weight: bold;
+    text-decoration: underline;
+    padding-bottom: 0.75em;
 }
 
-.saveTextsButton {
-  margin-left: 1em !important;
+.save-texts-button {
+    margin-left: 1em !important;
+    background-color: #9C27B0 !important;
+    font-weight: bold !important;
+    border-radius: 15px !important;
 }
 
 .accordionBorder {
-  border-radius: 0 0 4px 4px !important;
+    border-radius: 0 0 4px 4px !important;
 }
 
 .openedResultsAccordion {
-  margin-bottom: 1em !important;
+    margin-bottom: 1em !important;
 }
