+++ ui/src/elle/translations/en/translations.js
@@ -32,6 +32,7 @@ export const TRANSLATIONS_EN = {
   common_people: 'Employees',
   common_publications: 'Publications',
   common_publish_your_text: 'Publish your text',
+  common_query: 'Text query',
   common_search: 'Search',
   common_sentences: 'Sentence count',
   common_syllables: 'Syllables',
@@ -53,6 +54,7 @@ export const TRANSLATIONS_EN = {
   common_translation: 'Translation (Google Translate)',
   common_us: 'Us',
   common_view: 'View',
+  common_word_analyser: 'Word analyser',
   common_word_analysis: 'Word Analysis',
   common_word_in_context: 'Word in Context',
   common_wordlist: 'Wordlist',
@@ -69,6 +71,10 @@ export const TRANSLATIONS_EN = {
   concordances_sentences: 'sentences',
   concordances_view: 'View',
   concordances_words: 'words',
+  corpus_authors: 'Different authors',
+  corpus_donation_text: 'Help us research the Estonian langage by donating your texts to our corpus.',
+  corpus_texts: 'Texts in corpus',
+  donate_text: 'Donate your text',
   end: 'end',
   error_404_page_not_found: '404: page not found',
   error_file_upload_failed: 'File upload failed!',
