+++ ui/src/elle/translations/et/translations.js
@@ -32,6 +32,7 @@ export const TRANSLATIONS_ET = {
   common_people: 'Töötajad',
   common_publications: 'Üllitised',
   common_publish_your_text: 'Loovuta tekst',
+  common_query: 'Tekstipäring',
   common_search: 'Otsi',
   common_sentences: 'Lauseid',
   common_syllables: 'Silbid',
@@ -53,6 +54,7 @@ export const TRANSLATIONS_ET = {
   common_translation: 'Tõlge (Google Translate)',
   common_us: 'Meist',
   common_view: 'Kuva',
+  common_word_analyser: 'Sõnaanalüsaator',
   common_word_analysis: 'Sõnaanalüüs',
   common_word_in_context: 'Sõna kontekstis',
   common_wordlist: 'Sõnaloend',
@@ -69,6 +71,10 @@ export const TRANSLATIONS_ET = {
   concordances_sentences: 'lauset',
   concordances_view: 'Kuva',
   concordances_words: 'sõna',
+  corpus_authors: 'Erinevat autorit',
+  corpus_donation_text: 'Aita kaasa eesti keele uurimisele ning loovuta oma kirjateos meie korpusesse.',
+  corpus_texts: 'Teksti korpuses',
+  donate_text: 'Loovuta oma tekst',
   end: 'lõpp',
   error_404_page_not_found: '404: lehte ei leitud',
   error_file_upload_failed: 'Failide üleslaadimine ebaõnnestus!',
