+++ ui/src/elle/utils/constants.js
@@ -45,6 +45,50 @@ export const ButtonStyleSmall = {
   }
 }
 
+export const TabStyle = {
+  "& button:hover": {backgroundColor: "rgba(204, 168, 253, 1)", transition: "0.5s"},
+  "& button:focus": {
+    backgroundColor: "#9C27B0",
+    color: "black",
+    borderColor: "#9C27B0"
+  },
+  "& button:active": {backgroundColor: "#9C27B0"},
+  "& button": {
+    backgroundColor: "rgba(255, 208, 253, 1)",
+    color: "black",
+    fontWeight: "bold",
+    height: "60px",
+    borderBottom: "4px solid #9C27B0",
+    borderTop: "4px solid #9C27B0",
+    textTransform: "capitalize",
+    transition: "0.5s",
+  },
+  "& button:first-child": {
+    borderRadius: "25px 0 0 25px",
+    paddingLeft: "25px",
+    borderLeft: "4px solid #9C27B0"
+  },
+  "& button:last-child": {
+    borderRadius: "0 25px 25px 0",
+    paddingRight: "25px",
+    borderRight: "4px solid #9C27B0"
+  },
+  "& button.Mui-selected": {
+    backgroundColor: "#9C27B0",
+    color: "white",
+    borderColor: "rgba(204, 168, 253, 1)",
+    transition: "0.5s"
+  },
+  "& .MuiTabs-flexContainer": {
+    border: "none",
+    borderBottom: "none",
+  },
+  "& .MuiTabs-root": {
+    border: "none",
+    borderBottom: "none",
+  }
+}
+
 export const ElleOuterDivStyle = {
   background: 'linear-gradient(315deg, rgba(156, 39, 176, 1) 0%, rgba(156, 39, 176, 1) 33%, rgba(204, 168, 253, 1) 33%, rgba(204, 168, 253, 1) 67%, rgba(255, 208, 253, 1) 67%, rgba(255, 208, 253, 1) 100%)',
   height: '100%',
